move process chunking package
