+++ pkg/process/checks/chunking.go
@@ -7,94 +7,8 @@ package checks
 
 import (
 	model "github.com/DataDog/agent-payload/v5/process"
+	"github.com/DataDog/datadog-agent/pkg/process/util"
 )
-
-// payloadList is an abstract list of payloads subject to chunking
-type payloadList interface {
-	// Len returns the length of the list
-	Len() int
-	// WeightAt returns weight for the payload at position `idx` in the list
-	WeightAt(idx int) int
-	// ToChunk copies a slice from the list to an abstract connected chunker providing the cumulative weight of the chunk
-	ToChunk(start, end int, weight int)
-}
-
-// chunkAllocator abstracts management operations for chunk allocation
-type chunkAllocator interface {
-	// TakenSize returns the size allocated to the current chunk
-	TakenSize() int
-	// TakenWeight returns the weight allocated to the current chunk
-	TakenWeight() int
-	// Append creates a new chunk at the end (cases when it is known any previously allocated chunks cannot fit the payload)
-	Append()
-	// Next moves to the next chunk or allocates a new chunk if the current chunk is the last
-	Next()
-}
-
-// chunkPayloadsBySizeAndWeight allocates chunks of payloads taking max allowed size and max allowed weight
-// algorithm in the nutshell:
-// - iterate through payloads in the `payloadList`
-// - keep track of size and weight available for allocation (`TakenSize` and `TakenWeight`)
-// - create a new chunk once we exceed these limits
-// - consider case when the current item exceeds the max allowed weight and create a new chunk at the end (`Append`)
-// this implementation allows for multiple pases through the chunks, which can be useful in cases with different payload types
-// being allocated within chunks
-func chunkPayloadsBySizeAndWeight(l payloadList, a chunkAllocator, maxChunkSize int, maxChunkWeight int) {
-	start := 0
-	chunkWeight := 0
-	// Take available size and available weight by consulting the current chunk
-	availableSize := maxChunkSize - a.TakenSize()
-	availableWeight := maxChunkWeight - a.TakenWeight()
-	for i := 0; i < l.Len(); i++ {
-		itemWeight := l.WeightAt(i)
-		// Evaluate size of the currently accumulated items (from the start of the candidate chunk)
-		size := i - start
-		// Track if we need to skeep the item on the next chunk (large item chunked separately)
-		skipItem := false
-		if size >= availableSize || chunkWeight+itemWeight > availableWeight {
-			// We are exceeding available size or available weight and it is time to create a new chunk
-			if size > 0 {
-				// We already accumulated some items - create a new chunk
-				l.ToChunk(start, i, chunkWeight)
-				a.Next()
-			}
-			// Reset chunk weight
-			chunkWeight = 0
-			// Reset chunk start position
-			start = i
-			// Check if the current item exceeds the max allowed chunk weight
-			if itemWeight >= maxChunkWeight {
-				// Current item is exceeding max allowed chunk weight and should be chunked separately
-				if availableWeight < maxChunkWeight {
-					// Currently considered chunk already has allocations - create a new chunk at the end
-					a.Append()
-				}
-				// Chunk a single iem
-				l.ToChunk(i, i+1, itemWeight)
-				a.Next()
-				// Skip over this single item
-				start = i + 1
-				skipItem = true
-			} else {
-				// Find a chunk that can take the current items
-				for maxChunkSize-a.TakenSize() < 1 || maxChunkWeight-a.TakenWeight() < itemWeight {
-					a.Next()
-				}
-			}
-			// Reset available size and available weight based ont he current chunk
-			availableSize = maxChunkSize - a.TakenSize()
-			availableWeight = maxChunkWeight - a.TakenWeight()
-		}
-		if !skipItem {
-			// Only include the current item if it hasn't been to a separate chunk
-			chunkWeight += itemWeight
-		}
-	}
-	// Chunk the remainder of payloads
-	if start < l.Len() {
-		l.ToChunk(start, l.Len(), chunkWeight)
-	}
-}
 
 // chunkProcessesBySizeAndWeight chunks `model.Process` payloads by max allowed size and max allowed weight of a chunk
 func chunkProcessesBySizeAndWeight(procs []*model.Process, ctr *model.Container, maxChunkSize, maxChunkWeight int, chunker *collectorProcChunker) {
@@ -118,7 +32,7 @@ func chunkProcessesBySizeAndWeight(procs []*model.Process, ctr *model.Container,
 		procs:   procs,
 		chunker: chunker,
 	}
-	chunkPayloadsBySizeAndWeight(list, chunker, maxChunkSize, maxChunkWeight)
+	util.ChunkPayloadsBySizeAndWeight(list, chunker, maxChunkSize, maxChunkWeight)
 }
 
 // processList is a payload list of `model.Process` payloads
@@ -195,7 +109,7 @@ type collectorProcChunker struct {
 
 // collectprProcChunker implements both `chunkAllocator` and `processChunker`
 var _ processChunker = &collectorProcChunker{}
-var _ chunkAllocator = &collectorProcChunker{}
+var _ util.ChunkAllocator = &collectorProcChunker{}
 
 func (c *collectorProcChunker) Accept(procs []*model.Process, weight int) {
 	if c.idx >= len(c.collectorProcs) {
